requirements.txt
transformers==4.55.4        # :contentReference[oaicite:0]{index=0}
transformers[sentencepiece]==4.55.4
datasets==3.6.0              # :contentReference[oaicite:1]{index=1}
sacrebleu==2.5.1             # :contentReference[oaicite:2]{index=2}
rouge_score==0.1.2           # :contentReference[oaicite:3]{index=3}
py7zr==0.22.0                # :contentReference[oaicite:4]{index=4}
pandas==2.3.3                # :contentReference[oaicite:5]{index=5}
nltk                       # (latest version unspecified here)
tqdm                       # (latest version unspecified here)
PyYAML==6.0.3               # :contentReference[oaicite:6]{index=6}
matplotlib                 # (latest version unspecified here)
torch==2.9.0                # :contentReference[oaicite:7]{index=7}
notebook                   # (latest version unspecified here)
boto3                      # (latest version unspecified here)
mypy-boto3-s3==1.40.61      # :contentReference[oaicite:8]{index=8}
python-box==6.0.2           # (you specified this version)
typeguard==4.3.0               # :contentReference[oaicite:9]{index=9}
fastapi==0.119.0            # :contentReference[oaicite:10]{index=10}
uvicorn==0.18.3             # (you specified this version)
Jinja2==3.1.2               # (you specified this version)

-e .


